+++ packages/core/src/observer.ts
@@ -82,7 +82,7 @@ export class Observer {
     let root: Root;
 
     if (!Observer.interfaces || !Observer.interfaces.length) {
-      throw new Error('[Observer] Must define at least one adapter interface.');
+      throw new Error(`[Observer] Must define at least one adapter interface. "${JSON.stringify(Observer.interfaces)}"`);
     }
 
     if (!!~Observer.interfaces.indexOf(this.keypath[0])) {