+++ packages/core/src/vendors/index.ts
@@ -1,3 +1,2 @@
 export * from './debug.module';
-export * from './jquery.module';
 export * from './polyfills.module';