+++ packages/jquery/src/index.ts
@@ -1,11 +1,5 @@
-import { IRibaModule } from '@ribajs/core';
-
-import * as binders from './binders';
-
-export const i18nModule = (): IRibaModule => {
-  return {
-    binders,
-  };
-};
-
-export default i18nModule;
+export * from './vendors';
+export * from './binders';
+import { jqueryModule } from './jquery.module';
+export { jqueryModule };
+export default jqueryModule;