+++ packages/jquery/src/jquery.module.ts
@@ -1,0 +1,12 @@
+import { IRibaModule } from '@ribajs/core';
+import * as binders from './binders';
+// import * as formatters from './formatters';
+// import * as services from './services';
+// import * as components from './components';
+
+export const jqueryModule: IRibaModule = {
+  formatters: {},
+  binders,
+  services: {},
+  components: {},
+};